+++ noq-proto/src/tests/mod.rs
@@ -36,7 +36,7 @@ pub(crate) use util::*;
 mod encode_decode;
 mod multipath;
 #[cfg(not(all(target_family = "wasm", target_os = "unknown")))]
-mod proptest;
+mod proptests;
 #[cfg(not(all(target_family = "wasm", target_os = "unknown")))]
 mod random_interaction;
 mod token;
